+++ tools/prometheus-exporter.go
@@ -9,10 +9,12 @@ import (
 	"sync"
 	"time"
 
+	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 
 	"github.com/streamingfast/cli"
 	"github.com/streamingfast/cli/sflags"
+	"github.com/streamingfast/dgrpc"
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
@@ -52,7 +54,7 @@ func init() {
 	prometheusCmd.Flags().String("substreams-api-key-envvar", "SUBSTREAMS_API_KEY", "Name of variable containing Substreams Api Key")
 	prometheusCmd.Flags().BoolP("insecure", "k", false, "Skip certificate validation on GRPC connection")
 	prometheusCmd.Flags().BoolP("plaintext", "p", false, "Establish GRPC connection in plaintext")
-	prometheusCmd.Flags().Bool("force-v2", false, "Force using V2 API")
+	prometheusCmd.Flags().Int("force-protocol-version", 0, "Force the use of a specific protocol version (0=unset/auto, 2=v2, 3=v3)")
 	prometheusCmd.Flags().Int64("block-height", -1, "Block number to request (defaults to -1, which means the HEAD)")
 	prometheusCmd.Flags().Duration("max-freshness", time.Minute*2, "(only used if block-height is relative, i.e. below 0) check the age of the received blocks, fail an endpoint if it is older than this duration")
 	prometheusCmd.Flags().Duration("interval", time.Second*20, "endpoints will be polled at this interval")
@@ -173,7 +175,9 @@ func runPrometheus(cmd *cobra.Command, args []string) error {
 	plaintext := sflags.MustGetBool(cmd, "plaintext")
 	interval := sflags.MustGetDuration(cmd, "interval")
 	timeout := sflags.MustGetDuration(cmd, "timeout")
-	forceV2 := sflags.MustGetBool(cmd, "force-v2")
+	protocolVersionFlag := sflags.MustGetInt(cmd, "force-protocol-version")
+	forceProtocolVersion, err := client.ParseProtocolVersion(protocolVersionFlag)
+
 	maxFreshness := sflags.MustGetDuration(cmd, "max-freshness")
 
 	allLabels := map[string]bool{"endpoint": true}
@@ -207,15 +211,15 @@ func runPrometheus(cmd *cobra.Command, args []string) error {
 			startBlock = int64(*endpointMap[endpoint].startBlock)
 		}
 
-		substreamsClientConfig := client.NewSubstreamsClientConfig(
-			endpoint,
-			authToken,
-			authType,
-			insecure,
-			plaintext,
-			"substreams_prometheus",
-			forceV2,
-		)
+		substreamsClientConfig := client.NewSubstreamsClientConfig(client.SubstreamsClientConfigOptions{
+			Endpoint:             endpoint,
+			AuthToken:            authToken,
+			AuthType:             authType,
+			Insecure:             insecure,
+			PlainText:            plaintext,
+			Agent:                "substreams_prometheus",
+			ForceProtocolVersion: forceProtocolVersion,
+		})
 		var fresh *time.Duration
 		if startBlock < 0 {
 			fresh = &maxFreshness
@@ -281,7 +285,7 @@ func launchSubstreamsPoller(endpoint string, substreamsClientConfig *client.Subs
 			cancel()
 			continue
 		}
-		
+
 		ssClientV2 := pbsubstreamsrpcv2.NewStreamClient(conn)
 		ssClientV3 := pbsubstreamsrpcv3.NewStreamClient(conn)
 
@@ -310,8 +314,9 @@ func launchSubstreamsPoller(endpoint string, substreamsClientConfig *client.Subs
 		callOpts = append(callOpts, grpc.WaitForReady(false))
 		zlog.Debug("calling sf.substreams.rpc.v2.Stream/Blocks", zap.String("endpoint", endpoint), zap.String("output_module", outputStreamName), zap.Int64("start_block", blockNum), zap.Uint64("stop_block", stopBlockNum))
 
+		var isRunningV2 bool
 		var cli grpc.ServerStreamingClient[pbsubstreamsrpc.Response]
-		if substreamsClientConfig.ForceV2() {
+		if substreamsClientConfig.ForceProtocolVersion().IsV2() {
 			reqV2, err := subReq.ToV2()
 			if err != nil {
 				zlog.Error("call sf.substreams.rpc.v2.Stream/Blocks", zap.String("endpoint", endpoint), zap.Error(err))
@@ -328,14 +333,16 @@ func launchSubstreamsPoller(endpoint string, substreamsClientConfig *client.Subs
 				cancel()
 				continue
 			}
-		}
-		cli, err = ssClientV3.Blocks(ctx, subReq, callOpts...)
-		if err != nil {
-			zlog.Error("call sf.substreams.rpc.v2.Stream/Blocks", zap.String("endpoint", endpoint), zap.Error(err))
-			markFailure(endpoint, begin, err)
-			connClose()
-			cancel()
-			continue
+			isRunningV2 = true
+		} else {
+			cli, err = ssClientV3.Blocks(ctx, subReq, callOpts...)
+			if err != nil {
+				zlog.Error("call sf.substreams.rpc.v2.Stream/Blocks", zap.String("endpoint", endpoint), zap.Error(err))
+				markFailure(endpoint, begin, err)
+				connClose()
+				cancel()
+				continue
+			}
 		}
 
 	forloop:
@@ -371,6 +378,33 @@ func launchSubstreamsPoller(endpoint string, substreamsClientConfig *client.Subs
 				}
 			}
 			if err != nil {
+				if substreamsClientConfig.ForceProtocolVersion().IsUnset() && !isRunningV2 {
+					if dgrpcError := dgrpc.AsGRPCError(err); dgrpcError != nil {
+						switch dgrpcError.Code() {
+						case codes.Unimplemented, codes.NotFound:
+
+							zlog.Debug("server does not implement sf.substreams.rpc.v3.Stream/Blocks, trying v2")
+							reqV2, err := subReq.ToV2()
+							if err != nil {
+								zlog.Error("call sf.substreams.rpc.v2.Stream/Blocks", zap.String("endpoint", endpoint), zap.Error(err))
+								markFailure(endpoint, begin, err)
+								connClose()
+								cancel()
+								break
+							}
+							cli, err = ssClientV2.Blocks(ctx, reqV2, callOpts...)
+							if err != nil {
+								zlog.Error("call sf.substreams.rpc.v2.Stream/Blocks", zap.String("endpoint", endpoint), zap.Error(err))
+								markFailure(endpoint, begin, err)
+								connClose()
+								cancel()
+								continue
+							}
+							isRunningV2 = true
+						}
+					}
+				}
+
 				markFailure(endpoint, begin, err)
 				break
 			}